NATIVE-579: Refactor build config directory structure
