+++ build-configs/integreat/assets/app-logo.png
