+++ build-configs/integreat/theme/colors.js
@@ -1,7 +1,7 @@
 // @flow
 
-import { commonLightColors, commonDarkColors } from '../common/colors'
-import type { ColorsType } from '../common/colors'
+import { commonLightColors, commonDarkColors } from '../../common/colors'
+import type { ColorsType } from '../../common/colors'
 
 const themeColor = '#fbda16'
