+++ build-configs/integreat/theme/index.js
@@ -1,13 +1,13 @@
 // @flow
 
 import { lightColors, darkColors } from './colors'
-import malteFonts from './fonts'
-import dimensions from '../common/dimensions'
-import type { ThemeType } from '../ThemeType'
+import integreatFonts from './fonts'
+import dimensions from '../../common/dimensions'
+import type { ThemeType } from '../../ThemeType'
 
 export const lightTheme: ThemeType = {
   colors: lightColors,
-  fonts: malteFonts,
+  fonts: integreatFonts,
   dimensions
 }
