+++ build-configs/malte/theme/index.js
@@ -1,13 +1,13 @@
 // @flow
 
 import { lightColors, darkColors } from './colors'
-import integreatFonts from './fonts'
-import dimensions from '../common/dimensions'
-import type { ThemeType } from '../ThemeType'
+import malteFonts from './fonts'
+import dimensions from '../../common/dimensions'
+import type { ThemeType } from '../../ThemeType'
 
 export const lightTheme: ThemeType = {
   colors: lightColors,
-  fonts: integreatFonts,
+  fonts: malteFonts,
   dimensions
 }
