+++ src/modules/app/constants/buildConfig.js
@@ -1,20 +1,20 @@
 // @flow
 
-import type { BuildConfigType } from '../../../../build-configs/configs/BuildConfigType'
+import type { BuildConfigType } from '../../../../build-configs/BuildConfigType'
 import loadBuildConfig from '../../../../build-configs'
-import integreatAppLogo from '../../../../build-configs/assets/integreat/app-logo.png'
-import integreatLocationMarker from '../../../../build-configs/assets/integreat/location-marker.svg'
-import malteAppLogo from '../../../../build-configs/assets/malte/app-logo.svg'
-import malteLocationMarker from '../../../../build-configs/assets/malte/location-marker.svg'
-import integreatIntroLanguageIcon from '../../../../build-configs/assets/integreat/Language.svg'
-import integreatIntroEventsIcon from '../../../../build-configs/assets/integreat/Events.svg'
-import integreatIntroOffersIcon from '../../../../build-configs/assets/integreat/Offers.svg'
-import integreatIntroSearchIcon from '../../../../build-configs/assets/integreat/Search.svg'
-import malteIntroLanguageIcon from '../../../../build-configs/assets/malte/Language.svg'
-import malteIntroEventsIcon from '../../../../build-configs/assets/malte/Events.svg'
-import malteIntroOffersIcon from '../../../../build-configs/assets/malte/Offers.svg'
-import malteIntroSearchIcon from '../../../../build-configs/assets/malte/Search.svg'
-import { INTEGREAT_ASSETS, MALTE_ASSETS } from '../../../../build-configs/assets/AssetsType'
+import integreatAppLogo from '../../../../build-configs/integreat/assets/app-logo.png'
+import integreatLocationMarker from '../../../../build-configs/integreat/assets/location-marker.svg'
+import malteAppLogo from '../../../../build-configs/malte/assets/app-logo.svg'
+import malteLocationMarker from '../../../../build-configs/malte/assets/location-marker.svg'
+import integreatIntroLanguageIcon from '../../../../build-configs/integreat/assets/Language.svg'
+import integreatIntroEventsIcon from '../../../../build-configs/integreat/assets/Events.svg'
+import integreatIntroOffersIcon from '../../../../build-configs/integreat/assets/Offers.svg'
+import integreatIntroSearchIcon from '../../../../build-configs/integreat/assets/Search.svg'
+import malteIntroLanguageIcon from '../../../../build-configs/malte/assets/Language.svg'
+import malteIntroEventsIcon from '../../../../build-configs/malte/assets/Events.svg'
+import malteIntroOffersIcon from '../../../../build-configs/malte/assets/Offers.svg'
+import malteIntroSearchIcon from '../../../../build-configs/malte/assets/Search.svg'
+import { INTEGREAT_ASSETS, MALTE_ASSETS } from '../../../../build-configs/AssetsType'
 
 type IconSetType = {|
   appLogo: string,
